fixup! Rename GraphOperator to LinkedGraph
