+++ fedot/core/dag/graph_delegate.py
@@ -3,7 +3,7 @@ from typing import Union, Sequence, List, Optional, Tuple, Type
 
 from fedot.core.dag.graph import Graph
 from fedot.core.dag.graph_node import GraphNode
-from fedot.core.dag.graph_operator import LinkedGraph
+from fedot.core.dag.linked_graph import LinkedGraph
 
 
 class GraphDelegate(Graph):
