+++ fedot/core/pipelines/pipeline_builder.py
@@ -1,8 +1,7 @@
 from copy import deepcopy
 from typing import List, Union, Optional, Tuple, Dict
 
-from fedot.core.dag.graph import Graph
-from fedot.core.dag.graph_operator import LinkedGraph
+from fedot.core.dag.linked_graph import LinkedGraph
 from fedot.core.pipelines.node import Node, PrimaryNode, SecondaryNode
 from fedot.core.pipelines.pipeline import Pipeline
