+++ test/unit/visualization/test_graph_visualizations.py
@@ -6,7 +6,7 @@ import pytest
 from fedot.core.dag.graph import Graph
 from fedot.core.dag.graph_delegate import GraphDelegate
 from fedot.core.dag.linked_graph_node import LinkedGraphNode
-from fedot.core.dag.graph_operator import LinkedGraph
+from fedot.core.dag.linked_graph import LinkedGraph
 from fedot.core.optimisers.graph import OptGraph, OptNode
 from fedot.core.pipelines.node import Node
 from fedot.core.pipelines.pipeline import Pipeline
